Implement improved syntax.
Now we specify site usage with:

SITE USAGE [TOKEN $token] $path

This allows users to specify both a space token and a pathname;
tokens are likely easier to use with SRM-centric sites while the
pathname is likely easier to use with the classic quota case.

Both used and free space are included in the response; this prevents
us from having to do multiple queries for the same information.